Move #link_to into TagHelper
It seemed silly (and not particularly future-oriented) to have a module named _just_ so it could contain one method only.
